+++ Files/Changelog.md
@@ -3,6 +3,7 @@
 4.1.5: 01.05.2020
     -Refractored Shortcuts.js
     -Changed league Detection
+    -Fixed Prefix change not working
 
 4.1.4: 30.04.2020
     -Refractored league Detection