+++ Files/Changelog.md
@@ -4,6 +4,7 @@
     -Refractored Shortcuts.js
     -Changed league Detection
     -Fixed Prefix change not working
+    -Made WebsocketIp as global Object for easy swap
 
 4.1.4: 30.04.2020
     -Refractored league Detection